Extract ocSDKConfiguration to new file
