+++ src/config/ocSDKConfiguration.ts
@@ -1,0 +1,8 @@
+const ocSDKConfiguration = {
+    getChatTokenRetryCount: 2,
+    getChatTokenTimeBetweenRetriesOnFailure: 2000,
+    getChatTokenRetryOn429: false,
+    useUnauthReconnectIdSigQueryParam: false
+};
+
+export default ocSDKConfiguration;
